@@ -22,13 +22,14 @@
     },
     "dependencies": {
         "@skbkontur/react-icons": "4.1.0",
-        "@skbkontur/react-stack-layout": "^1.0.3",
-        "@skbkontur/react-ui": "2.0.0",
+        "@skbkontur/react-stack-layout": "1.0.3",
+        "@skbkontur/react-ui": "2.4.0",
         "@skbkontur/react-ui-validations": "1.1.3",
         "bundle-loader": "0.5.6",
         "color-hash": "1.0.3",
         "date-fns": "2.9.0",
         "dompurify": "^1.0.10",
+        "file-saver": "2.0.2",
         "husky": "4.2.3",
         "jest": "24.5.0",
         "lodash": "4.17.13",
@@ -58,7 +59,7 @@
         "@storybook/addons": "^5.3.17",
         "@storybook/react": "^5.3.17",
         "aws-sdk": "^2.649.0",
-        "babel-eslint": "10.0.1",
+        "babel-eslint": "10.1.0",
         "babel-loader": "8.0.5",
         "babel-plugin-module-resolver": "^4.0.0",
         "classnames": "2.2.6",
@@ -66,16 +67,16 @@
         "creevey": "0.5.7",
         "cross-env": "5.2.0",
         "css-loader": "2.1.1",
-        "eslint": "5.15.3",
-        "eslint-config-airbnb": "17.1.0",
-        "eslint-config-prettier": "4.1.0",
+        "eslint": "7.2.0",
+        "eslint-config-airbnb": "18.1.0",
+        "eslint-config-prettier": "6.11.0",
         "eslint-plugin-babel": "5.3.0",
         "eslint-plugin-flowtype": "3.4.2",
-        "eslint-plugin-import": "2.16.0",
-        "eslint-plugin-jsx-a11y": "6.2.1",
-        "eslint-plugin-prettier": "3.0.1",
-        "eslint-plugin-react": "7.12.4",
-        "eslint-plugin-react-hooks": "^2.5.0",
+        "eslint-plugin-import": "2.21.1",
+        "eslint-plugin-jsx-a11y": "6.2.3",
+        "eslint-plugin-prettier": "3.1.3",
+        "eslint-plugin-react": "7.20.0",
+        "eslint-plugin-react-hooks": "4.0.4",
         "file-loader": "3.0.1",
         "flow-bin": "0.95.1",
         "flow-coverage-report": "0.6.1",
