@@ -2,12 +2,15 @@
 import * as React from "react";
 import { Button } from "@skbkontur/react-ui/components/Button";
 import { Modal } from "@skbkontur/react-ui/components/Modal";
-import { Gapped } from "@skbkontur/react-ui/components/Gapped";
 import { ValidationContainer } from "@skbkontur/react-ui-validations";
+import { Fill, RowStack } from "@skbkontur/react-stack-layout";
 import type { Contact } from "../../Domain/Contact";
+import { omitSubscription } from "../../helpers/omitTypes";
 import SubscriptionEditor, {
     type SubscriptionInfo,
 } from "../SubscriptionEditor/SubscriptionEditor";
+import FileLoader from "../FileLoader/FileLoader";
+import ModalError from "../ModalError/ModalError";
 
 type Props = {
     subscription: SubscriptionInfo,
@@ -22,6 +25,7 @@ type Props = {
 type State = {
     createInProcess: boolean,
     createAndTestInProcess: boolean,
+    error?: string,
 };
 
 export default class SubscriptionEditModal extends React.Component<Props, State> {
@@ -39,8 +43,8 @@ export default class SubscriptionEditModal extends React.Component<Props, State>
     }
 
     render(): React.Node {
-        const { subscription, tags, contacts, onChange, onCancel } = this.props;
-        const { createInProcess, createAndTestInProcess } = this.state;
+        const { subscription, tags, contacts, onCancel } = this.props;
+        const { createInProcess, createAndTestInProcess, error } = this.state;
         const isActionButtonsDisabled = createInProcess || createAndTestInProcess;
 
         return (
@@ -50,38 +54,43 @@ export default class SubscriptionEditModal extends React.Component<Props, State>
                     <ValidationContainer ref={this.validationContainer}>
                         <SubscriptionEditor
                             subscription={subscription}
-                            onChange={onChange}
+                            onChange={this.handleChange}
                             tags={tags}
                             contacts={contacts}
                         />
                     </ValidationContainer>
                 </Modal.Body>
                 <Modal.Footer panel sticky>
-                    <Gapped gap={10}>
+                    <ModalError message={error} maxWidth="450px" />
+                    <RowStack gap={2} block baseline>
                         <Button
                             use="primary"
                             disabled={isActionButtonsDisabled}
                             loading={createInProcess}
-                            onClick={() => {
-                                this.handleCreate();
-                            }}
+                            onClick={this.handleCreate}
                         >
                             Add
                         </Button>
                         <Button
                             disabled={isActionButtonsDisabled}
                             loading={createAndTestInProcess}
-                            onClick={() => {
-                                this.handleCreateAndTest();
-                            }}
+                            onClick={this.handleCreateAndTest}
                         >
                             Add and test
                         </Button>
-                    </Gapped>
+                        <Fill />
+                        <FileLoader onLoad={this.handleImport}>Import subscription</FileLoader>
+                    </RowStack>
                 </Modal.Footer>
             </Modal>
         );
     }
+
+    handleChange = (subscription: $Shape<SubscriptionInfo>) => {
+        const { onChange } = this.props;
+        onChange(subscription);
+        this.setState({ error: undefined });
+    };
 
     handleCreate = async () => {
         if (!(await this.validateForm())) {
@@ -106,6 +115,22 @@ export default class SubscriptionEditModal extends React.Component<Props, State>
             await onCreateAndTestSubscription(subscription);
         } finally {
             this.setState({ createAndTestInProcess: false });
+        }
+    };
+
+    handleImport = (fileData: string, fileName: string) => {
+        try {
+            const subscription = JSON.parse(fileData);
+
+            if (typeof subscription === "object" && subscription !== null) {
+                this.handleChange(omitSubscription(subscription));
+            } else {
+                throw new Error(`Must be a subscription object`);
+            }
+        } catch (e) {
+            this.setState({
+                error: `File ${fileName} cannot be converted to subscription. ${e.message}`,
+            });
         }
     };
 
